Renames folder containing STL mesh readers to readers
In preparation for adding other readers, e.g. C2C.
